Add atuhentication and authorization

+++ src/Presentation/LibraryApp.API/Controllers/UserController.cs
@@ -1,4 +1,5 @@
-﻿using LibraryApp.Application.Common.Pagination;
+﻿using LibraryApp.API.Authorization;
+using LibraryApp.Application.Common.Pagination;
 using LibraryApp.Application.Feauters.Books.Querries.Dto;
 using LibraryApp.Application.Feauters.Books.Querries.GetUserReadBooks;
 using LibraryApp.Application.Feauters.Reviews.Queries.Dto;
@@ -14,7 +15,9 @@ using LibraryApp.Application.Feauters.Users.Queries.Dto;
 using LibraryApp.Application.Feauters.Users.Queries.GetUserAvatar;
 using LibraryApp.Application.Feauters.Users.Queries.GetUserDetails;
 using LibraryApp.Application.Feauters.Users.Queries.GetUsers;
+using LibraryApp.Application.Feauters.Users.Queries.Login;
 using MediatR;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace LibraryApp.API.Controllers
@@ -24,10 +27,30 @@ namespace LibraryApp.API.Controllers
 	public class UserController : ControllerBase
 	{
 		private readonly IMediator _mediator;
+		private readonly IAuthorizationService _authorizationService;
 
-		public UserController(IMediator mediator)
+		public UserController(IMediator mediator, IAuthorizationService authorizationService)
 		{
 			_mediator = mediator;
+			_authorizationService = authorizationService;
+		}
+
+		[HttpGet("login")]
+		public async Task<ActionResult<LoginResponse>> Login(
+			LoginQuery command, CancellationToken cancellationToken)
+		{
+			var response = await _mediator.Send(command, cancellationToken);
+
+			return Ok(response);
+		}
+
+		[HttpPost]
+		public async Task<ActionResult<Guid>> Register(
+			[FromBody] CreateUserCommand command, CancellationToken cancellationToken)
+		{
+			var response = await _mediator.Send(command, cancellationToken);
+
+			return CreatedAtAction(nameof(Register), response);
 		}
 
 		[HttpGet]
@@ -41,24 +64,23 @@ namespace LibraryApp.API.Controllers
 			return Ok(response);
 		}
 
-		[HttpPost]
-		public async Task<ActionResult<Guid>> Create(
-			[FromBody] CreateUserCommand command, CancellationToken cancellationToken)
-		{
-			var response = await _mediator.Send(command, cancellationToken);
-
-			return CreatedAtAction(nameof(Create), response);
-		}
-
 		[HttpPut]
+		[Authorize]
 		public async Task<ActionResult> Update(
 			[FromBody] UpdateUserCommand command, CancellationToken cancellationToken)
 		{
+			var authorizationResult = await _authorizationService
+				.AuthorizeAsync(User, command.UserId, Policies.UserUpdatePolicyName);
+
+			if (authorizationResult.Succeeded == false) 
+				return new ForbidResult();
+
 			await _mediator.Send(command, cancellationToken);
 
 			return NoContent();
 		}
 
+	
 		[HttpGet("{id}")]
 		public async Task<ActionResult<UserDetailsDto>> GetById(
 			Guid id, CancellationToken cancellationToken)
@@ -70,8 +92,15 @@ namespace LibraryApp.API.Controllers
 		}
 
 		[HttpDelete("{id}")]
+		[Authorize]
 		public async Task<ActionResult> Delete(Guid id, CancellationToken cancellationToken)
 		{
+			var authorizationResult = await _authorizationService
+				.AuthorizeAsync(User, id, Policies.UserDeletePolicyName);
+
+			if (authorizationResult.Succeeded == false)
+				return new ForbidResult();
+
 			var command = new DeleteUserCommand(id);
 			await _mediator.Send(command, cancellationToken);
 
@@ -88,18 +117,33 @@ namespace LibraryApp.API.Controllers
 		}
 
 		[HttpPut("{id}/avatar")]
+		[Authorize]
 		public async Task<ActionResult> UpdateAvatar(
-			[FromForm] UpdateUserAvatarCommand command, CancellationToken cancellationToken)
+			Guid id, [FromForm] IFormFile avatarFile, CancellationToken cancellationToken)
 		{
+			var authorizationResult = await _authorizationService
+				.AuthorizeAsync(User, id, Policies.UserUpdatePolicyName);
+
+			if (authorizationResult.Succeeded == false)
+				return new ForbidResult();
+
+			var command = new UpdateUserAvatarCommand(id, avatarFile);
 			await _mediator.Send(command, cancellationToken);
 
 			return NoContent();
 		}
 
 		[HttpDelete("{id}/avatar")]
+		[Authorize]
 		public async Task<ActionResult> DeleteAvatar(
 			Guid id, CancellationToken cancellationToken)
 		{
+			var authorizationResult = await _authorizationService
+				.AuthorizeAsync(User, id, Policies.UserDeletePolicyName);
+
+			if (authorizationResult.Succeeded == false)
+				return new ForbidResult();
+
 			var command = new DeleteUserAvatarCommand(id);
 			await _mediator.Send(command, cancellationToken);
 
@@ -127,9 +171,16 @@ namespace LibraryApp.API.Controllers
 		}
 
 		[HttpPost("{userId}/read-books/{bookId}")]
+		[Authorize]
 		public async Task<ActionResult> AddReadBook(
 			Guid userId, Guid bookId, CancellationToken cancellationToken)
 		{
+			var authorizationResult = await _authorizationService
+				.AuthorizeAsync(User, userId, Policies.UserUpdatePolicyName);
+
+			if (authorizationResult.Succeeded == false)
+				return new ForbidResult();
+
 			var command = new AddReadBookCommand(userId, bookId);
 			await _mediator.Send(command, cancellationToken);
 
@@ -137,9 +188,16 @@ namespace LibraryApp.API.Controllers
 		}
 
 		[HttpDelete("{userId}/read-books/{bookId}")]
+		[Authorize]
 		public async Task<ActionResult> DeleteReadedBook(
 			Guid userId, Guid bookId, CancellationToken cancellationToken)
 		{
+			var authorizationResult = await _authorizationService
+				.AuthorizeAsync(User, userId, Policies.UserUpdatePolicyName);
+
+			if (authorizationResult.Succeeded == false)
+				return new ForbidResult();
+
 			var command = new DeleteReadBookCommand(userId, bookId);
 			await _mediator.Send(command, cancellationToken);
